fix: replace private with '#' for private members
With typescripts private access modifier, it can restrict  until we compile the code to js. Once the code compiled to js, all the private modifiers are striped and the private members are exposed to public. This can be fixed by replacing it with javascript's '#' modifier which is supported from typescript 3.8 onwards.